+++ src/main/java/com/example/spring/model/AutoComplete/Region.java
@@ -15,7 +15,6 @@ public class Region {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+    private Integer id;
     private String name;
 }
